Consider direction in encoder's estimate (#41)
The encoder class's estimateSpeed function was not considering direction in
it's RPM calculation, resulting in it always reporting a positive
value. An overloaded version of the function was added to consider the
sign of the target speed in order to properly report the sign of the
motor velocity.

The motor controller seems to have an issue where its speed will not
properly set to 0 if it is reporting a negative speed. This may be
related to this issue. Further investigation will be conducted and a new
issue may be opened.